Add video validation and dynamic project settings on export
Introduce ffprobe-based video validation and make export settings dynamic.
Added VideoValidator (uses ffprobe JSON output) in `dvre/utils/helper.py` to assert `width`/`height`/`frame-rate`;
README updated to document ffprobe requirement and FFPROBE_PATH.
The builder now constructs and passes the validator to MediaService;
MediaService.import_media signature updated to accept a validator and will validate media before importing.

ProjectService.export_project now accepts width, height and frame_rate and uses them for render settings.
Also changed TimelineSettings.frame_rate to float and updated example_config.json accordingly.

+++ dvre/utils/config.py
@@ -31,7 +31,7 @@ class TimelineSettings(BaseModel):
 
     width: int = Field(1920, gt=0, description="Timeline width in pixels")
     height: int = Field(1080, gt=0, description="Timeline height in pixels")
-    frame_rate: int = Field(60, gt=0, description="Frame rate (fps)")
+    frame_rate: float = Field(60, gt=0, description="Frame rate (fps)")
 
 
 class FusionClip(BaseModel):
